AI-Parking-Management-System-main/requirements.txt
Flask==2.3.3
supabase==2.0.2
speechrecognition==3.10.0
pyaudio==0.2.11
gtts==2.4.0
python-dotenv==1.0.0
pydub==0.25.1
